8363889: Update sun.print.PrintJob2D to use Disposer
